fix: admin panel QoL 2
